Add version and byte order fields to the pwanls file
The version is added to ensure that no old files are processed in case
the binary format is changed in the future.

The byte order field is used to verify that the byte order on the
system that produced the pwalns file is the same than on the system
that uses it.