Compute shortest path thoroughly
Earlier, the shortest path search aborted for a path P1 to a species
SpA in case SpA was already reached by a different path P2 with a
higher score. However, P1 most probably reaches SpA at a different
coord than P2. And paths to a subsequent species SpB from the coord of
P1 might result in a better score than reaching SpB from the coord of
P2.

This new implementation of Ipp::projectCoord() takes the above into
account and only aborts the search for a path P1 in case a different
path P2 reached the same species at the same coord with a higer score.